Integrate with live color correction
