+++ app-frontend/src/app/components/colorCorrectAdjust/colorCorrectAdjust.controller.js
@@ -4,45 +4,72 @@ export default class ColorCorrectAdjustController {
     }
 
     $onInit() {
-        let baseFilterOpts = {
-            floor: -100,
-            ceil: 100,
+        let baseGammaOptions = {
+            floor: 0,
+            ceil: 2,
+            step: 0.1,
+            precision: 1,
             onEnd: (id, val) => this.onFilterChange(id, val)
         };
-        this.redOpts = Object.assign({}, baseFilterOpts, {id: 'red'});
-        this.greenOpts = Object.assign({}, baseFilterOpts, {id: 'green'});
-        this.blueOpts = Object.assign({}, baseFilterOpts, {id: 'blue'});
-        this.brightOpts = Object.assign({}, baseFilterOpts, {id: 'brightness'});
-        this.ctrstOpts = Object.assign({}, baseFilterOpts, {id: 'contrast'});
 
-        this.zeroesCorrection = {
-            red: 0,
-            green: 0,
-            blue: 0,
-            brightness: 0,
-            contrast: 0
+        let baseFilterOptions = {
+            floor: 0,
+            ceil: 60,
+            step: 1,
+            onEnd: (id, val) => this.onFilterChange(id, val)
         };
-        // On init, the desired correction, if it exists, becomes the correction that
-        // we reset to.
-        if (this.desiredCorrection) {
-            // Allow external users to pass in just the attributes they want.
-            this.correction = Object.assign({}, this.zeroesCorrection, this.desiredCorrection);
-            this.resetCorrection = Object.assign({}, this.correction);
-        } else {
-            this.correction = Object.assign({}, this.zeroesCorrection);
-            this.resetCorrection = Object.assign({}, this.correction);
-        }
+
+        let alphaOptions = {
+            floor: 0,
+            disabled: true,
+            ceil: 1,
+            step: 0.1,
+            precision: 2,
+            onEnd: (id, val) => this.onFilterChange(id, val)
+        };
+
+        let betaOptions = {
+            floor: 0,
+            disabled: true,
+            ceil: 50,
+            step: 1,
+            onEnd: (id, val) => this.onFilterChange(id, val)
+        };
+
+        let minMaxOptions = {
+            floor: 0,
+            ceil: 20000,
+            step: 10,
+            onEnd: (id, val) => this.onFilterChange(id, val)
+        };
+
+        this.redGammaOptions = Object.assign({}, baseGammaOptions, {id: 'red'});
+        this.greenGammaOptions = Object.assign({}, baseGammaOptions, {id: 'green'});
+        this.blueGammaOptions = Object.assign({}, baseGammaOptions, {id: 'blue'});
+
+        this.brightnessOptions = Object.assign({}, baseFilterOptions, {id: 'brightness'});
+        this.contrastOptions = Object.assign({}, baseFilterOptions, {id: 'contrast'});
+
+        this.alphaOptions = Object.assign({}, alphaOptions, {id: 'alpha'});
+        this.betaOptions = Object.assign({}, betaOptions, {id: 'beta'});
+        this.minMaxOptions = Object.assign({}, minMaxOptions, {id: 'minmax'});
     }
 
     $onChanges(changesObj) {
-        if ('reset' in changesObj && this.resetCorrection) {
-            this.correction = Object.assign({}, this.resetCorrection);
-            this.onCorrectionChange({newCorrection: Object.assign({}, this.correction)});
+        if ('correction' in changesObj) {
+            this.myCorrection = Object.assign({}, changesObj.correction.currentValue);
         }
     }
 
+    /**
+     * Makes color correction changes available as a component output
+     *
+     * @param {string} id used to identify correction that has been modified
+     * @param {number} val new value for a color correction
+     * @returns {null} null
+     */
     onFilterChange(id, val) {
-        this.correction[id] = val;
-        this.onCorrectionChange({newCorrection: Object.assign({}, this.correction)});
+        this.myCorrection[id] = val;
+        this.onCorrectionChange({newCorrection: Object.assign({}, this.myCorrection)});
     }
 }
