+++ app-frontend/src/app/components/colorCorrectPane/colorCorrectPane.component.js
@@ -4,11 +4,7 @@ const colorCorrectPane = {
     templateUrl: colorCorrectPaneTpl,
     controller: 'ColorCorrectPaneController',
     bindings: {
-        sceneList: '<?',
-        // Pass in a boolean to toggle to reset filters
-        reset: '<?',
-        onCorrectionChange: '&',
-        selectedLayers: '<?'
+        selectedLayers: '<'
     }
 };
