+++ app-frontend/src/app/components/colorCorrectPane/colorCorrectPane.controller.js
@@ -8,77 +8,66 @@ export default class ColorCorrectPaneController {
         this.$q = $q;
     }
 
-
     $onInit() {
-        let baseGammaOptions = {
-            floor: 0,
-            ceil: 2,
-            step: 0.1,
-            precision: 1,
-            onEnd: (id, val) => this.onFilterChange(id, val)
-        };
-
-        let baseFilterOptions = {
-            floor: 0,
-            ceil: 60,
-            step: 1,
-            onEnd: (id, val) => this.onFilterChange(id, val)
-        };
-
-        let alphaOptions = {
-            floor: 0,
-            disabled: true,
-            ceil: 1,
-            step: 0.1,
-            precision: 2,
-            onEnd: (id, val) => this.onFilterChange(id, val)
-        };
-
-        let betaOptions = {
-            floor: 0,
-            disabled: true,
-            ceil: 50,
-            step: 1,
-            onEnd: (id, val) => this.onFilterChange(id, val)
-        };
-
-        let minMaxOptions = {
-            floor: 0,
-            ceil: 20000,
-            step: 10,
-            onEnd: (id, val) => this.onFilterChange(id, val)
-        };
-
-        this.redGammaOptions = Object.assign({}, baseGammaOptions, {id: 'redGamma'});
-        this.greenGammaOptions = Object.assign({}, baseGammaOptions, {id: 'greenGamma'});
-        this.blueGammaOptions = Object.assign({}, baseGammaOptions, {id: 'blueGamma'});
-
-        this.brightnessOptions = Object.assign({}, baseFilterOptions, {id: 'brightness'});
-        this.contrastOptions = Object.assign({}, baseFilterOptions, {id: 'contrast'});
-
-        this.alphaOptions = Object.assign({}, alphaOptions, {id: 'alpha'});
-        this.betaOptions = Object.assign({}, betaOptions, {id: 'beta'});
-        this.minMaxOptions = Object.assign({}, minMaxOptions, {id: 'minmax'});
-
+        // If the user navigates directly here we don't know what scenes to
+        // adjust so redirect to the scenes list.
+        if (this.selectedLayers.size === 0) {
+            this.$state.go('^.scenes');
+            return;
+        }
         // Initialize correction to first selected layer (if there are multiple)
+        this.correction = this.selectedLayers.values().next().value.baseColorCorrection();
+
+        // Fake data for our histogram; this will get replaced by a service call later.
+        this.red = [];
+        this.green = [];
+        this.blue = [];
+        for (let i of [0, 100, 200, 300, 400]) {
+            this.red.push({x: i, y: i});
+            this.green.push({x: i, y: 400 - i});
+            this.blue.push({x: i, y: i / 2.0});
+        }
+        this.data = [
+            {
+                values: this.red,
+                key: 'Red channel',
+                color: '#bb0000',
+                area: true
+            },
+            {
+                values: this.green,
+                key: 'Green channel',
+                color: '#00bb00',
+                area: true
+            },
+            {
+                values: this.blue,
+                key: 'Blue channel',
+                color: '#0000dd',
+                area: true
+            }
+        ];
+    }
+
+    resetCorrection() {
+        for (let layer of this.selectedLayers.values()) {
+            layer.resetTiles();
+        }
         this.correction = this.selectedLayers.values().next().value.baseColorCorrection();
     }
 
     /**
-     * Triggered when filters are changed
+     * Triggered when the adjustment pane reports changes to color correction
      *
      * Applies color corrections to all selected layers
-     * @param {string} id used to identify correction that has been modified
-     * @param {number} val new value for a color correction
+     * @param {object} newCorrection object to apply to each layer
      *
      * @returns {null} null
      */
-    onFilterChange(id, val) {
-        this.correction[id] = val;
-        let correction = this.correction;
-        this.selectedLayers.forEach(function (layer) {
-            layer.colorCorrect(correction);
-        });
-        this.onCorrectionChange({newCorrection: Object.assign({}, this.correction)});
+    onCorrectionChange(newCorrection) {
+        this.correction = newCorrection;
+        for (let layer of this.selectedLayers.values()) {
+            layer.colorCorrect(this.correction);
+        }
     }
 }
