+++ app-frontend/src/app/components/colorCorrectPane/colorCorrectPane.html
@@ -5,87 +5,9 @@
       <a ng-click="">Make source histogram</a>
     </div>
   </div>
-  <!-- Color correction -->
-  <h6>Gamma Correction</h6>
-  <div class="filter">
-    <label>Red</label>
-    <div class="filter-item slider-filter">
-      <div id="red_slider"></div>
-    </div>
-    <rzslider rz-slider-model="$ctrl.correction.red"
-              rz-slider-options="$ctrl.redGammaOptions"
-    ></rzslider>
-  </div>
-  <div class="filter">
-    <label>Green</label>
-    <div class="filter-item slider-filter">
-      <div id="green_slider"></div>
-    </div>
-    <rzslider rz-slider-model="$ctrl.correction.green"
-              rz-slider-options="$ctrl.greenGammaOptions"
-    ></rzslider>
-  </div>
-  <div class="filter">
-    <label>Blue</label>
-    <div class="filter-item slider-filter">
-      <div id="blue_slider"></div>
-    </div>
-    <rzslider rz-slider-model="$ctrl.correction.blue"
-              rz-slider-options="$ctrl.blueGammaOptions"
-    ></rzslider>
-  </div>
-
-    <!--TODO: Add Sigmoidal Back When Fixed ISSUE: 761-->
-    <!--<h6>Sigmoidal Correction</h6>-->
-    <!--<div class="filter">-->
-      <!--<label>Alpha</label>-->
-      <!--<div class="filter-item slider-filter">-->
-        <!--<div id="alpha_slider"></div>-->
-      <!--</div>-->
-      <!--<rzslider rz-slider-model="$ctrl.correction.alpha"-->
-                <!--rz-slider-options="$ctrl.alphaOptions"-->
-      <!--&gt;</rzslider>-->
-    <!--</div>-->
-    <!--<div class="filter">-->
-      <!--<label>Beta</label>-->
-      <!--<div class="filter-item slider-filter">-->
-        <!--<div id="beta_slider"></div>-->
-      <!--</div>-->
-      <!--<rzslider rz-slider-model="$ctrl.correction.beta"-->
-                <!--rz-slider-options="$ctrl.betaOptions"-->
-      <!--&gt;</rzslider>-->
-    <!--</div>-->
-
-  <div class="bc-filters-container">
-    <h6>Brightness/Contrast</h6>
-    <div class="filter">
-      <label>Brightness</label>
-      <rzslider rz-slider-model="$ctrl.correction.brightness"
-                rz-slider-options="$ctrl.brightnessOptions"></rzslider>
-    </div>
-    <div class="filter">
-      <label>Contrast</label>
-      <div class="filter-item slider-filter">
-        <div id="contrast"></div>
-      </div>
-      <rzslider rz-slider-model="$ctrl.correction.contrast"
-                rz-slider-options="$ctrl.contrastOptions"></rzslider>
-    </div>
-    <div class="filter">
-      <label>Min/Max Clipping</label>
-      <div class="filter-item slider-filter">
-        <div id="minmax"></div>
-      </div>
-      <rzslider rz-slider-model="$ctrl.correction.min"
-                rz-slider-high="$ctrl.correction.max"
-                rz-slider-options="$ctrl.minMaxOptions"
-      ></rzslider>
-    </div>
-  </div>
-
   <div class="sidebar-scrollable">
     <rf-color-correct-adjust
-        desired-correction="$ctrl.correction"
+        correction="$ctrl.correction"
         on-correction-change="$ctrl.onCorrectionChange(newCorrection)"
         reset="$ctrl.resetToggle"
         class="color-correct-panel">
@@ -94,7 +16,7 @@
   <div class="sidebar-static">
     <a href ng-click="$ctrl.$state.go('^.scenes')">Back</a>
     <div class="sidebar-actions">
-      <a href ng-click="$ctrl.resetColors()">Reset</a>
+      <a href ng-click="$ctrl.resetCorrection()">Reset</a>
       <a href ng-click="$ctrl.applyColors()">Apply changes</a>
     </div>
   </div>
