+++ app-frontend/src/app/components/colorCorrectPane/colorCorrectPane.state.html
@@ -1,0 +1,2 @@
+<rf-color-correct-pane selected-layers="$ctrl.selectedLayers">
+</rf-color-correct-pane>
