+++ app-frontend/src/app/components/colorCorrectScenes/colorCorrectScenes.component.js
@@ -4,10 +4,11 @@ const colorCorrectScenes = {
     templateUrl: colorCorrectScenesTpl,
     controller: 'ColorCorrectScenesController',
     bindings: {
-        sceneList: '<?',
-        // Pass in a boolean to toggle to reset filters
-        reset: '<?',
-        onCorrectionChange: '&'
+        selectedScenes: '=',
+        selectedLayers: '=',
+        sceneList: '=',
+        sceneLayers: '=',
+        layers: '='
     }
 };
