+++ app-frontend/src/app/components/colorCorrectScenes/colorCorrectScenes.state.html
@@ -1,0 +1,9 @@
+<rf-color-correct-scenes
+  selected-scenes="$ctrl.selectedScenes"
+  selected-layers="$ctrl.selectedLayers"
+  scene-list="$ctrl.sceneList"
+  scene-layers="$ctrl.sceneLayers"
+  layers="$ctrl.layers"
+  on-layer-change="$ctrl.updateLayers({layers: layers})">
+</rf-color-correct-scenes>
+
