+++ app-frontend/src/app/components/mosaicScenes/mosaicScenes.component.js
@@ -4,10 +4,8 @@ const mosaicScenes = {
     templateUrl: mosaicScenesTpl,
     controller: 'MosaicScenesController',
     bindings: {
-        sceneList: '<?',
-        // Pass in a boolean to toggle to reset filters
-        reset: '<?',
-        onCorrectionChange: '&'
+        sceneList: '=',
+        layers: '='
     }
 };
