+++ app-frontend/src/app/components/mosaicScenes/mosaicScenes.state.html
@@ -1,0 +1,4 @@
+<rf-mosaic-scenes 
+  scene-list="$ctrl.sceneList"
+  layers="$ctrl.layers">
+</rf-mosaic-scenes>
