+++ app-frontend/src/app/pages/editor/project/projectEdit.html
@@ -7,95 +7,6 @@
 </div>
 <div class="container column-stretch">
   <!-- Sidebar (container is in parent view) -->
-  <div class="sidebar scene-list-pane" ng-if="$ctrl.showLayerList" >
-    <!-- scene listing -->
-    <div class="sidebar-static">
-        <div class="btn-group" uib-dropdown dropdown-append-to-body>
-          <a href uib-dropdown-toggle>
-            Color mode <i class="icon-caret-down"></i>
-          </a>
-          <ul class="dropdown-menu" uib-dropdown-menu role="menu">
-            <li role="menuitem" ng-click="$ctrl.setBands('natural')">Natural color</li>
-            <li role="menuitem" ng-click="$ctrl.setBands('cir')">Color Infrared</li>
-            <li role="menuitem" ng-click="$ctrl.setBands('urban')">Urban</li>
-            <li role="menuitem" ng-click="$ctrl.setBands('water')">Water</li>
-            <li role="menuitem" ng-click="$ctrl.setBands('atmosphere')">Atmosphere Penetration</li>
-            <li role="menuitem" ng-click="$ctrl.setBands('agriculture')">Agriculture</li>
-            <li role="menuitem" ng-click="$ctrl.setBands('forestfire')">Forest Fire</li>
-            <li role="menuitem" ng-click="$ctrl.setBands('bareearth')">Bare Earth Change Detection</li>
-            <li role="menuitem" ng-click="$ctrl.setBands('vegwater')">Vegetation & Water</li>
-          </ul>
-        </div>
-        <div class="btn-group" uib-dropdown dropdown-append-to-body>
-          <a href uib-dropdown-toggle>
-            Action <i class="icon-caret-down"></i>
-          </a>
-          <ul class="dropdown-menu" uib-dropdown-menu role="menu">
-            <li role="menuitem"><i class="icon-load"></i> Sync histogram</li>
-            <li role="menuitem"><i class="icon-arrows-cw"></i> Reset changes</li>
-            <li role="menuitem"><i class="icon-trash"></i> Delete</li>
-          </ul>
-          <button ng-disabled="$ctrl.selectedScenes.size === 0" ng-click="$ctrl.openColorCorrect()">Color Correct</button>
-        </div>
-        <button class="btn btn-square"
-                ng-class="{'btn-secondary': !$ctrl.shouldSelectAll()}"
-                ng-click="$ctrl.onToggleSelection()">
-          <i ng-class="{'icon-plus': $ctrl.shouldSelectAll(),
-                        'icon-cross': !$ctrl.shouldSelectAll()}"></i>
-        </button>
-    </div>
-    <div class="sidebar-scrollable" id="infscroll">
-      <div class="list-group">
-        <rf-scene-item
-            ng-click="$ctrl.openColorCorrect(scene)"
-            scene="scene"
-            selectable
-            selected="$ctrl.isSelected(scene)"
-            on-select="$ctrl.setSelected(scene, selected)"
-            class="selectable"
-            ng-repeat="scene in $ctrl.sceneList track by scene.id">
-        </rf-scene-item>
-      </div>
-      <div class="list-group" ng-show="$ctrl.loading">
-        <span class="list-placeholder">
-          <i class="icon-load"></i>
-        </span>
-      </div>
-      <div class="list-group" ng-show="$ctrl.errorMsg">
-        <span class="list-placeholder">
-          {{$ctrl.errorMsg}}
-        </span>
-      </div>
-    </div>
-  </div>
-  <div class="sidebar sidebar-extended color-correct-pane"
-       ng-if="$ctrl.showColorCorrect">
-    <!-- Color correction -->
-    <div class="sidebar-static histogram">
-      <rf-channel-histogram data="$ctrl.data"></rf-channel-histogram>
-      <div>
-        <a ng-click="">Make source histogram</a>
-      </div>
-    </div>
-    <div class="sidebar-scrollable">
-      <rf-color-correct-pane
-         layer="$ctrl.selectedScene"
-         selected-layers="$ctrl.selectedLayers"
-         desired-correction="$ctrl.correction"
-         on-correction-change="$ctrl.onCorrectionChange(newCorrection)"
-         reset="$ctrl.resetToggle"
-         class="color-correct-panel">
-      </rf-color-correct-pane>
-    </div>
-    <!--TODO: This static part of the sidebar should move into the correction component ISSUE: 725-->
-    <div class="sidebar-static">
-      <a href ng-click="$ctrl.hideColorCorrect()">Back</a>
-      <div class="sidebar-actions">
-        <a href ng-click="$ctrl.resetColors()">Reset</a>
-      </div>
-    </div>
-  </div>
-
   <ui-view class="sidebar"></ui-view>
   <!-- Map -->
   <div class="main">
